Re-organizing auth unit tests
